Refactor code to follow coding standard
