+++ src/main/java/missminutes/Deadline.java
@@ -1,4 +1,4 @@
-package MissMinutes;
+package missminutes;
 
 import java.io.Serializable;
 import java.time.format.DateTimeParseException;
@@ -8,7 +8,7 @@ import java.time.LocalDateTime;
 
 public class Deadline extends Task implements Serializable {
     protected final LocalDateTime by;
-    protected final static Pattern formatter = Pattern.compile("(.+) /by (.+)");
+    protected final static Pattern FORMATTER = Pattern.compile("(.+) /by (.+)");
 
     public Deadline(String name, LocalDateTime by) {
         super(name);
@@ -16,14 +16,14 @@ public class Deadline extends Task implements Serializable {
     }
 
     public static Deadline fromStr(String input) throws MissMinutesException {
-        Matcher matcher = Deadline.formatter.matcher(input);
+        Matcher matcher = Deadline.FORMATTER.matcher(input);
 
         if (matcher.find()) {
             String name = matcher.group(1);
             String byStr = matcher.group(2);
 
             try {
-                LocalDateTime by = LocalDateTime.parse(byStr, Task.inputFormat);
+                LocalDateTime by = LocalDateTime.parse(byStr, Task.INPUT_DATE_FORMAT);
                 return new Deadline(name, by);
             } catch (DateTimeParseException err) {
                 throw new MissMinutesException("Please enter a valid date time format. For example, 2019-12-31 1800");
@@ -35,6 +35,6 @@ public class Deadline extends Task implements Serializable {
 
     @Override
     public String toString() {
-        return "[D]" + super.toString() + " (by: " + by.format(Task.outputFormat) + ")";
+        return "[D]" + super.toString() + " (by: " + by.format(Task.OUTPUT_DATE_FORMAT) + ")";
     }
 }
