+++ src/main/java/missminutes/Event.java
@@ -1,4 +1,4 @@
-package MissMinutes;
+package missminutes;
 
 import java.io.Serializable;
 import java.time.format.DateTimeParseException;
@@ -9,7 +9,7 @@ import java.time.LocalDateTime;
 public class Event extends Task implements Serializable {
     protected final LocalDateTime startTime;
     protected final LocalDateTime endTime;
-    protected static final Pattern formatter = Pattern.compile("(.+) /from (.+) /to (.+)");
+    protected static final Pattern FORMATTER = Pattern.compile("(.+) /from (.+) /to (.+)");
 
     public Event(String name, LocalDateTime startTime, LocalDateTime endTime) {
         super(name);
@@ -18,7 +18,7 @@ public class Event extends Task implements Serializable {
     }
 
     public static Event fromStr(String input) throws MissMinutesException {
-        Matcher matcher = Event.formatter.matcher(input);
+        Matcher matcher = Event.FORMATTER.matcher(input);
 
         if (matcher.find()) {
             String name = matcher.group(1);
@@ -27,8 +27,8 @@ public class Event extends Task implements Serializable {
 
             LocalDateTime startTime, endTime;
             try {
-                startTime = LocalDateTime.parse(startTimeStr, Task.inputFormat);
-                endTime = LocalDateTime.parse(endTimeStr, Task.inputFormat);
+                startTime = LocalDateTime.parse(startTimeStr, Task.INPUT_DATE_FORMAT);
+                endTime = LocalDateTime.parse(endTimeStr, Task.INPUT_DATE_FORMAT);
             } catch (DateTimeParseException err) {
                 throw new MissMinutesException("Please enter a valid date time format. For example, 2019-12-31 1800");
             }
@@ -41,7 +41,7 @@ public class Event extends Task implements Serializable {
 
     @Override
     public String toString() {
-        return "[E]" + super.toString() + " (from: " + startTime.format(Task.outputFormat)
-                + " to: " + endTime.format(Task.outputFormat) + ")";
+        return "[E]" + super.toString() + " (from: " + startTime.format(Task.OUTPUT_DATE_FORMAT)
+                + " to: " + endTime.format(Task.OUTPUT_DATE_FORMAT) + ")";
     }
 }
