+++ src/main/java/missminutes/Storage.java
@@ -1,4 +1,4 @@
-package MissMinutes;
+package missminutes;
 
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
