+++ src/main/java/missminutes/Task.java
@@ -1,4 +1,4 @@
-package MissMinutes;
+package missminutes;
 
 import java.io.Serializable;
 import java.time.format.DateTimeFormatter;
@@ -7,8 +7,8 @@ public class Task implements Serializable {
     protected String description;
     protected boolean isDone;
 
-    protected static final DateTimeFormatter inputFormat = DateTimeFormatter.ofPattern("yyyy-MM-dd HHmm");
-    protected static final DateTimeFormatter outputFormat = DateTimeFormatter.ofPattern("MMM d yyyy ha");
+    protected static final DateTimeFormatter INPUT_DATE_FORMAT = DateTimeFormatter.ofPattern("yyyy-MM-dd HHmm");
+    protected static final DateTimeFormatter OUTPUT_DATE_FORMAT = DateTimeFormatter.ofPattern("MMM d yyyy ha");
 
     public Task(String description) {
         this.description = description;
