+++ src/main/java/missminutes/Ui.java
@@ -1,4 +1,4 @@
-package MissMinutes;
+package missminutes;
 
 import java.util.ArrayList;
 import java.util.Scanner;
