+++ src/test/java/missminutes/DeadlineTest.java
@@ -1,4 +1,4 @@
-package MissMinutes;
+package missminutes;
 
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.assertEquals;
