+++ text-ui-test/runtest.bat
@@ -8,7 +8,7 @@ if exist ACTUAL.TXT del ACTUAL.TXT
 if exist tasks.bin del tasks.bin
 
 REM compile the code into the bin folder
-javac  -cp ..\src\main\java\MissMinutes -Xlint:none -d ..\bin ..\src\main\java\MissMinutes\*.java
+javac  -cp ..\src\main\java\missminutes -Xlint:none -d ..\bin ..\src\main\java\missminutes\*.java
 IF ERRORLEVEL 1 (
     echo ********** BUILD FAILURE **********
     exit /b 1
@@ -16,7 +16,7 @@ IF ERRORLEVEL 1 (
 REM no error here, errorlevel == 0
 
 REM run the program, feed commands from input.txt file and redirect the output to the ACTUAL.TXT
-java -classpath ..\bin MissMinutes.MissMinutes < input.txt > ACTUAL.TXT
+java -classpath ..\bin missminutes.missminutes < input.txt > ACTUAL.TXT
 
 REM compare the output to the expected output
 FC ACTUAL.TXT EXPECTED.TXT
